Move WarehouseItems related pages to upper folder
